requirements.txt
agno
openai
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq